requirements.txt
streamlit==1.31.0
langchain==0.2.0
langchain-community==0.2.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
requests==2.31.0  # Added for custom xAI API calls